Modularize server
